requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dotmap==1.3.30
enum-compat==0.0.3
frozenlist==1.3.3
idna==3.4
meshtastic==2.1.8
multidict==6.0.4
paho-mqtt==1.6.1
pexpect==4.8.0
protobuf==4.23.3
ptyprocess==0.7.0
py-cord==2.4.1
pygatt==4.0.5
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0
requests==2.31.0
tabulate==0.9.0
timeago==1.0.16
urllib3==2.0.3
yarl==1.9.2
